+++ src/main/java/com/honiism/discord/lemi/commands/slash/currency/Inventory.java
@@ -22,7 +22,7 @@ package com.honiism.discord.lemi.commands.slash.currency;
 import com.honiism.discord.lemi.commands.handler.CommandCategory;
 import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
-import com.honiism.discord.lemi.data.UserDataManager;
+import com.honiism.discord.lemi.data.currency.UserDataManager;
 
 import java.util.HashMap;
 import java.util.concurrent.TimeUnit;
