+++ src/main/java/com/honiism/discord/lemi/commands/text/handler/TextCmd.java
@@ -22,7 +22,7 @@ package com.honiism.discord.lemi.commands.text.handler;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.honiism.discord.lemi.commands.handler.CommandCategory;
 import com.honiism.discord.lemi.commands.handler.UserCategory;
-import com.honiism.discord.lemi.data.UserDataManager;
+import com.honiism.discord.lemi.data.currency.UserDataManager;
 import com.honiism.discord.lemi.data.database.managers.LemiDbBalManager;
 import com.honiism.discord.lemi.utils.misc.Tools;
