+++ src/main/java/com/honiism/discord/lemi/data/currency/UserDataManager.java
@@ -17,7 +17,7 @@
  * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.honiism.discord.lemi.data;
+package com.honiism.discord.lemi.data.currency;
 
 import java.util.ArrayList;
 import java.util.List;
