+++ src/main/java/com/honiism/discord/lemi/data/database/LemiDbBalDs.java
@@ -31,9 +31,9 @@ import java.util.ArrayList;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.honiism.discord.lemi.Lemi;
-import com.honiism.discord.lemi.data.InventoryData;
-import com.honiism.discord.lemi.data.UserData;
-import com.honiism.discord.lemi.data.UserDataManager;
+import com.honiism.discord.lemi.data.currency.InventoryData;
+import com.honiism.discord.lemi.data.currency.UserData;
+import com.honiism.discord.lemi.data.currency.UserDataManager;
 import com.honiism.discord.lemi.data.database.managers.LemiDbBalManager;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
@@ -146,14 +146,14 @@ public class LemiDbBalDs implements LemiDbBalManager {
                 PreparedStatement insertStatement =
                     conn.prepareStatement("INSERT INTO user_currency_data(user_id, json_data) VALUES(?, ?)")) {
 
-            UserData userImpl = new UserData(userId);
+            UserData userData = new UserData(userId);
 
-            userImpl.setBalance(1000);
-            userImpl.setdeaths(0);
-            userImpl.setPassiveMode(false);
-            userImpl.setInventory(new ArrayList<InventoryData>());
+            userData.setBalance(1000);
+            userData.setdeaths(0);
+            userData.setPassiveMode(false);
+            userData.setInventory(new ArrayList<InventoryData>());
 
-            String jsonData = Lemi.getInstance().getObjectMapper().writeValueAsString(userImpl);
+            String jsonData = Lemi.getInstance().getObjectMapper().writeValueAsString(userData);
             
             insertStatement.setLong(1, userId);
             insertStatement.setString(2, jsonData);
